fix: simplify AlbyResponseError message generation, add test
